refactor home screen view model and interactions
