+++ app/src/main/java/com/baghdad/tudee/ui/screens/homeScreen/HomeScreen.kt
@@ -46,7 +46,6 @@ import com.baghdad.tudee.ui.composable.taskDetailsBottomSheet.TaskDetailsBottomS
 import com.baghdad.tudee.ui.screens.homeScreen.addEditTask.AddEditTaskBottomSheet
 import com.baghdad.tudee.ui.utils.formatDate
 import com.baghdad.tudee.ui.utils.now
-import com.baghdad.tudee.viewModel.homescreenViewModel.HomeScreenViewModel
 import kotlinx.datetime.LocalDate
 import org.koin.androidx.compose.koinViewModel
